chore: move detector to its own file
