+++ packages/opentelemetry-node/package.json
@@ -124,7 +124,6 @@
     "@opentelemetry/semantic-conventions": "^1.30.0",
     "@opentelemetry/winston-transport": "^0.11.0",
     "import-in-the-middle": "^1.12.0",
-    "json-bigint": "^1.0.0",
     "safe-stable-stringify": "^2.4.3"
   },
   "devDependencies": {
